refactor: set payload to maximum bytes we can detect
Confirmed UART flow control connection working
Confirmed 252 bytes data transmission limit (at least by the SDK).
Anything over 252 bytes  will cause an infinite loop in the Bluetooth SDK
and anything over 182 bytes cannot be received by other bluetooth devices.
This is perhaps hardware limitation. Still need to confirm.

See line 960 of sli_bt_api.h and examine
the struct at line 3137.

+++ projects/stm32f411_freertos/l4_io/silabs_ble/silabs_ble_freertos.c
@@ -136,7 +136,7 @@ void silabs_ble_freertos__initialize(unsigned long priority)
 
 bool silabs_ble_freertos__send_notification(uint8_t connection,
                                             uint16_t characteristic_handle,
-                                            uint32_t length,
+                                            uint8_t length,
                                             const uint8_t *values)
 {
   sl_status_t status;
