+++ projects/stm32f411_freertos/l4_io/silabs_ble/silabs_ble_freertos.h
@@ -26,7 +26,7 @@ void silabs_ble_freertos__initialize(unsigned long priority);
 
 bool silabs_ble_freertos__send_notification(uint8_t connection,
                                             uint16_t characteristic_handle,
-                                            uint32_t length,
+                                            uint8_t length,
                                             const uint8_t *values);
 
 bool silabs_ble_freertos__register_on_event_callback(silabs_ble_freertos__on_event_callback_f on_event_cb);
